Add Dockerfile and docker publishing steps (#25)
Adds a Dockerfile and an Actions job to publish a container based on
that Dockerfile. Additionally, introduces toggles to the publish
workflow to opt in/out of publishing to pypi or docker (these should
usually be updated in sync, but this gives us an escape hatch for fixing
problematic release runs or testing workflow changes).